requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
nltk
validators
sqlalchemy
python-dateutil
PyYAML